Simplify looking up user defaults from params namespace
This commit simplifies the process by which the default
value of $user is determined when not provided.

The memcached class inherits from the params class so
that the params class defaults are accessible as
parameter defaults.

It is worth noting that this code will not work with versions
2.6.0 -> 2.6.5

It is also worth noting that this will probably just be a
temporary refactor until hiera becomes more pervasive.