Make sysconfig configuration work with undef value for logfile
With the setting of data types for parameters, the logfile is now of type
Optional[Stdlib::Absolutepath]. This means that we can no longer set the
logfile to be an empty string, and thus we memcached configuration
doesn't do a relevant check for it.

Instead, we now handle undef, which is a valid value with the Optional
type.