Remove checking for $listen_ip
Stdlib::compat::Ip_address is only checking for a single IP Address, whereas memcached can actually listen for an array of IP Addresses that is comma separated. Thus the memcached module is articially limiting a valid configuration option in memcached.
I don't have a better idea for how to validate, but this Pull Request at least removes the problem.